rename package to cpisdk
