+++ pkg/cpisdk/common_system_test.go
@@ -2,7 +2,7 @@
    Copyright 2021 VMware, Inc.
    SPDX-License-Identifier: Apache-2.0
 */
-package vcdcpiclient
+package cpisdk
 
 import (
 	"fmt"
