+++ pkg/cpisdk/cpi_gateway_manager.go
@@ -1,4 +1,4 @@
-package vcdcpiclient
+package cpisdk
 
 import (
 	"context"
