+++ pkg/cpisdk/cpi_gateway_manager_system_test.go
@@ -1,4 +1,4 @@
-package vcdcpiclient
+package cpisdk
 
 import (
 	"context"
