+++ pkg/cpisdk/rde.go
@@ -1,4 +1,4 @@
-package vcdcpiclient
+package cpisdk
 
 import (
 	"context"
