+++ pkg/cpisdk/rde_test.go
@@ -1,9 +1,8 @@
-package vcdcpiclient
+package cpisdk
 
 import (
 	"context"
 	"github.com/stretchr/testify/assert"
-	"github.com/vmware/cloud-provider-for-cloud-director/pkg/vcdsdk"
 	"io/ioutil"
 	"k8s.io/apimachinery/pkg/util/yaml"
 	"net/http"
@@ -23,18 +22,18 @@ func foundStringInSlice(find string, slice []string) bool {
 func TestUpdateRDEUsingEtag(t *testing.T) {
 	// TODO: This test will currently fail unless the code below is uncommented. Refer to VCDA-3600
 
-	cloudConfig, err := vcdsdk.GetTestConfig()
+	cloudConfig, err := getTestConfig()
 	assert.NoError(t, err, "There should be no error opening and parsing cloud config file contents.")
 
-	authFile := filepath.Join(GitRoot, "testdata/auth_test.yaml")
+	authFile := filepath.Join(gitRoot, "testdata/auth_test.yaml")
 	authFileContent, err := ioutil.ReadFile(authFile)
 	assert.NoError(t, err, "There should be no error reading the auth file contents.")
 
-	var authDetails vcdsdk.AuthorizationDetails
+	var authDetails authorizationDetails
 	err = yaml.Unmarshal(authFileContent, &authDetails)
 	assert.NoError(t, err, "There should be no error parsing auth file content.")
 
-	vcdClient, err := vcdsdk.GetTestVCDClient(cloudConfig, map[string]interface{}{
+	vcdClient, err := getTestVCDClient(cloudConfig, map[string]interface{}{
 		"user":    authDetails.Username,
 		"secret":  authDetails.Password,
 		"userOrg": authDetails.UserOrg,
